requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
black==22.1.0
blue==0.9.1
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
colorama==0.4.6
coverage==7.4.1
cryptography==42.0.5
dnspython==2.5.0
ecdsa==0.18.0
email-validator==2.1.0.post1
factory-boy==3.3.0
Faker==24.4.0
fastapi==0.109.0
flake8==4.0.1
freezegun==1.4.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
Mako==1.3.2
MarkupSafe==2.1.5
mccabe==0.6.1
mypy-extensions==1.0.0
packaging==23.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.1.0
pluggy==1.4.0
psutil==5.9.8
pyasn1==0.5.1
pycodestyle==2.8.0
pycparser==2.21
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pyflakes==2.4.0
pytest==7.4.4
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
taskipy==1.12.2
tomli==2.0.1
typing_extensions==4.9.0
uvicorn==0.27.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
